requirements.txt
openai
docx2txt
langchain
langchain_experimental
pinecone-client
python-dotenv
pypdf
wikipedia
tiktoken